Do not produce output to the screen.
From a user perspective, it is very difficult to tell where this information
was produced, and so is not very useful. It's also obscures legitimate
output in projects that use SampleFlow if they want to generate screen
output themselves. As a consequence, just do what the function
documentation says: Return a default-constructed object.