Remove some commentary.
These comments explain what *doesn't* work, but that is not useful information
to the reader: The reader wants to understand what *does* happen in a given
file. It's also information that tends to go out of date pretty quickly: If
we implement the ability to work with things other than valarray, we would
have to update comments in files that are unrelated to what we are
implementing.